pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=20.0,<21.0",
]
